requirements.txt
pywebview
gTTS
genanki
eng-to-ipa
deep-translator